feat: 개인정보, 필수입력 영역 컴포넌트 분리
